feat(blue-migration-sdk-viem): rename package
