+++ packages/blue-migration-sdk-viem/src/fetchers/aaveV2/aaveV2.helpers.ts
