+++ packages/blue-migration-sdk-viem/src/fetchers/aaveV3Optimizer/AaveV3.maths.ts
