+++ packages/blue-migration-sdk-viem/src/fetchers/aaveV3Optimizer/aaveV3Optimizer.helpers.ts
