+++ packages/blue-migration-sdk-viem/src/fetchers/compoundV2/compoundV2.fetchers.ts
@@ -18,11 +18,11 @@ import {
 } from "../../types/index.js";
 import { rateToApy } from "../../utils/rates.js";
 
-import { fetchAccruedExchangeRate } from "./compoundV2.helpers.js";
 import type { DeploylessFetchParameters } from "@morpho-org/blue-sdk-viem";
 import type { Client } from "viem";
 import { getChainId, readContract } from "viem/actions";
 import { cErc20Abi, cEtherAbi, mErc20Abi } from "../../abis/compoundV2.abis.js";
+import { fetchAccruedExchangeRate } from "./compoundV2.helpers.js";
 
 const COMPOUNDING_PERIOD: { [chainID in ChainId]: Time.PeriodLike } = {
   [ChainId.BaseMainnet]: "s",
