+++ packages/blue-migration-sdk-viem/src/fetchers/compoundV3/compoundV3.fetchers.ts
@@ -1,6 +1,10 @@
 import { type Address, ChainUtils } from "@morpho-org/blue-sdk";
 import { isDefined, values } from "@morpho-org/morpho-ts";
 
+import type { DeploylessFetchParameters } from "@morpho-org/blue-sdk-viem";
+import type { Client } from "viem";
+import { getChainId, readContract } from "viem/actions";
+import { cometAbi, cometExtAbi } from "../../abis/compoundV3.abis.js";
 import MIGRATION_ADDRESSES from "../../config.js";
 import type { MigratablePosition } from "../../positions/index.js";
 import { MigratableSupplyPosition_CompoundV3 } from "../../positions/supply/compoundV3.supply.js";
@@ -9,10 +13,6 @@ import {
   SupplyMigrationLimiter,
 } from "../../types/index.js";
 import { rateToApy } from "../../utils/rates.js";
-import type { DeploylessFetchParameters } from "@morpho-org/blue-sdk-viem";
-import { getChainId, readContract } from "viem/actions";
-import type { Client } from "viem";
-import { cometAbi, cometExtAbi } from "../../abis/compoundV3.abis.js";
 
 async function fetchCompoundV3InstancePosition(
   user: Address,
