+++ packages/blue-migration-sdk-viem/src/positions/borrow/blue.borrow.ts
@@ -1,11 +1,11 @@
 import {
-  type MarketId,
   type ChainId,
-  type Market,
-  type Position,
   DEFAULT_SLIPPAGE_TOLERANCE,
-  getChainAddresses,
+  type Market,
+  type MarketId,
   MathLib,
+  type Position,
+  getChainAddresses,
 } from "@morpho-org/blue-sdk";
 import type { BlueInputBundlerOperations } from "@morpho-org/bundler-sdk-viem";
 import { maxUint256 } from "viem";
