+++ packages/blue-migration-sdk-viem/src/positions/supply/aaveV3.supply.ts
@@ -16,15 +16,13 @@ import {
   SupplyMigrationLimiter,
 } from "../../types/index.js";
 
-import {
-  MigratableSupplyPosition,
-  type MigratableSupplyPositionConfig,
-} from "./index.js";
+import { getPermitTypedData } from "@morpho-org/blue-sdk-viem";
 import type {
-  SignatureRequirement,
   Action,
+  SignatureRequirement,
 } from "@morpho-org/bundler-sdk-viem";
-import { signTypedData } from "viem/actions";
+import BundlerAction from "@morpho-org/bundler-sdk-viem/src/BundlerAction.js";
+import { baseBundlerAbi } from "@morpho-org/bundler-sdk-viem/src/abis.js";
 import {
   type Account,
   type Client,
@@ -32,10 +30,12 @@ import {
   maxUint256,
   verifyTypedData,
 } from "viem";
-import { getPermitTypedData } from "@morpho-org/blue-sdk-viem";
+import { signTypedData } from "viem/actions";
 import { aTokenV3Abi } from "../../abis/aaveV3.abis.js";
-import { baseBundlerAbi } from "@morpho-org/bundler-sdk-viem/src/abis.js";
-import BundlerAction from "@morpho-org/bundler-sdk-viem/src/BundlerAction.js";
+import {
+  MigratableSupplyPosition,
+  type MigratableSupplyPositionConfig,
+} from "./index.js";
 
 interface MigratableSupplyPositionConfig_AaveV3
   extends Omit<MigratableSupplyPositionConfig, "protocol"> {
