--- packages/migration-sdk-viem/README.md
@@ -1,34 +1,0 @@
-# @morpho-org/migration-sdk-viem
-
-<a href="https://www.npmjs.com/package/@morpho-org/migration-sdk-viem">
-    <picture>
-        <source media="(prefers-color-scheme: dark)" srcset="https://img.shields.io/npm/v/@morpho-org/migration-sdk-viem?colorA=21262d&colorB=21262d&style=flat">
-        <img src="https://img.shields.io/npm/v/@morpho-org/migration-sdk-viem?colorA=f6f8fa&colorB=f6f8fa&style=flat" alt="Version">
-    </picture>
-</a>
-<a href="https://github.com/morpho-org/migration-sdk-viem/blob/main/LICENSE">
-    <picture>
-        <source media="(prefers-color-scheme: dark)" srcset="https://img.shields.io/npm/l/@morpho-org/migration-sdk-viem?colorA=21262d&colorB=21262d&style=flat">
-        <img src="https://img.shields.io/npm/l/@morpho-org/migration-sdk-viem?colorA=f6f8fa&colorB=f6f8fa&style=flat" alt="MIT License">
-    </picture>
-</a>
-<a href="https://www.npmjs.com/package/@morpho-org/migration-sdk-viem">
-    <picture>
-        <source media="(prefers-color-scheme: dark)" srcset="https://img.shields.io/npm/dm/@morpho-org/migration-sdk-viem?colorA=21262d&colorB=21262d&style=flat">
-        <img src="https://img.shields.io/npm/dm/@morpho-org/migration-sdk-viem?colorA=f6f8fa&colorB=f6f8fa&style=flat" alt="Downloads per month">
-    </picture>
-</a>
-<br />
-<br />
-
-Viem-based extension of `@morpho-org/simulation-sdk` that exports utilities to build migration bundles to migrate lending positions (on aave, compound, morpho-aaveV3-optimizer, ...) to the morpho protocol.
-
-## Installation
-
-```bash
-npm install @morpho-org/migration-sdk-viem
-```
-
-```bash
-yarn add @morpho-org/migration-sdk-viem
-```
